+++ web/cypress/support/commands.js
@@ -24,4 +24,4 @@
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
 
-// import 'cypress-file-upload';
+import 'cypress-file-upload';
