+++ web/cypress/support/index.js
@@ -15,7 +15,6 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands';
-import 'cypress-file-upload';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
