+++ web/src/components/CreateTestPlugins/Postman/steps/UploadCollection/fields/SelectTestFromCollection.tsx
@@ -1,7 +1,7 @@
 import {Form, FormInstance, Select} from 'antd';
 import {useWatch} from 'antd/es/form/Form';
 import React, {Dispatch, SetStateAction} from 'react';
-import {RequestDefinitionExtended} from '../hooks/getRequestsFromCollection';
+import {RequestDefinitionExtended} from 'services/PostmanService.service';
 import {useSelectTestCallback} from '../hooks/useSelectTestCallback';
 import {IRequestDetailsValues} from '../UploadCollection';
