--- web/src/components/CreateTestPlugins/Postman/steps/UploadCollection/hooks/substituteVariable.tsx
@@ -1,13 +1,0 @@
-import {VariableDefinition} from 'postman-collection';
-
-export function substituteVariable(variables: VariableDefinition[], value: string | undefined) {
-  const regExpMatchArray = (value || '').match(/\{{([^}]+)}}/);
-  return regExpMatchArray
-    ? value
-        ?.replaceAll(
-          regExpMatchArray?.[0],
-          variables.find(variable => variable.key === regExpMatchArray?.[1])?.value || value
-        )
-        .replaceAll(' ', '')
-    : value;
-}
